[CP-3166] `string-to-regex` moved to standalone catalog
