+++ libs/generic-view/utils/src/index.ts
@@ -13,5 +13,5 @@ export * from "./lib/models/modal.types"
 export * from "./lib/data-provider-helpers"
 export * from "./lib/data-sort"
 export * from "./lib/forms-provider/forms-provider"
-export * from "./lib/data-provider-helpers/string-to-regex"
+export * from "./lib/string-to-regex/string-to-regex"
 export * from "./lib/use-current-view-key"
