+++ libs/generic-view/utils/src/lib/string-to-regex/string-to-regex.test.ts
